cdc-streaming/consumers-msk/shared/requirements.txt
confluent-kafka>=2.3.0
aws-msk-iam-sasl-signer-python>=1.0.0
backports.zoneinfo>=0.2.1; python_version < "3.9"

